Avoid infinite loop on some OS/device combinations
The system uses the intrinsic content size of the scroll view to add width/height constraints to the scroll view. In this process, the system calls -updateConstraintsIfNeeded on the scroll view. However, since -invalidateIntrinsicContentSize also marks the constraints as needing updates, when the system calls -updateConstraintsIfNeeded, it's updating more constraints than it expects, causing layout of the entire view to happen again. We can avoid this by calling -updateConstraintsIfNeeded manually so that when the system calls this method, only the width and height constraints are actually updated.